Add precise control of feedback in validationInput
A new function, validationInputWithFeedback is created, which returns,
beside the validationInput, the error feedback as a separate widget of
type `m ()`.

This is done to enable more precise control when positioning html
elements, so that for example a dropdown menu can be attached directly
to the input element. The original `validationInput` function has been
rewritten in terms of the new one, but the semantic is unchanged.