Revert back to the more baked-in scheme for tasks
The scheme that encodes more checkout logic in SQL can be elaborated
on in the future to allow less latency in the checkout phase of the
job by reducing roundtrips with the database. Ideally the entire
checkout logic can be encoded in a single statement which can be
commited before the result is acknowledged by the client.