Add lower level helpers to notify-listen-beam
Namely:
  * runPgInsertReturningListWithNotify
  * runPgUpdateReturningListWithNotify
  * runPgDeleteReturningListWithNotify

The lower level helpers do not assume anything about the statement that
will be executed. Instead, they take as a parameter the function for
converting the result of the statement to a list of notifications.

Some of the higher level helpers which construct their own queries have
been reimplemented using the lower level helpers and more efficient
returns.

I have taken the opportunity of removing a constraint synonym for a pile
of beam noise by using the constraint synonyms beam itself uses. This
ought to be a little more useful for figuring out how to fulfill these
constraints when they aren't automatically satisfied.