Fix current_timestamp_
The type of the Postgres expression `current_timestamp at time zone 'UTC'` is `TIMESTAMP WITHOUT TIME ZONE`. Consequently, if we attempt to store this expression in a column of type `TIMESTAMP WITH TIME ZONE` then Postgres will cast the `TIMESTAMP WITHOUT TIME ZONE`-value into a `TIMESTAMP WITH TIME ZONE` *using the time zone configured for the Postgres instance in question*. This results in `current_timestamp at time zone 'UTC'` being wrong for all Postgres instances running with a non-UTC time zone (ie. most).

Note that Postgres stores a value of type `TIMESTAMP WITH TIME ZONE` as a UTC timestamp. Ie. it does _not_ store the time zone component anywhere -- a timestamp with a time zone component is simply converted to UTC and then stored. So when `current_timestamp` returns e.g. `2022-08-08 11:59:35.958693+02`, Postgres will correctly convert this to the UTC time `2022-08-08 09:59:35.958693` -- whereas storing `2022-08-08 09:59:35.958693` in the same column would be cast to UTC+2 (`2022-08-08 09:59:35.958693+02`) and thus be two hours too early.