Put all rhyolite packages in the shell
This means that when you drop into `nix-shell -A proj.shells.ghc`, cabal
will build all of the local packages that are dependencies you have,
instead of them being built as part of entering the nix-shell. This is
somewhat painful in that more building happens than needed for most
dependencies, but the previous approach led to being unable to enter a
nix-shell if there was a broken package.