move obelisk thunk so that rhyolite doesn't appear as an obelisk project
the '.obelisk' path causds `ob run` and similar to commands to deduce that
rhyolite "is an obelisk project", and won't interpret packages in ghci when
rhyolite is used as an unpacked nix-thunk dependency.  this happens even
when the user passes '--interpret path/to/rhyolte' on the command line.

rhyolite isn't an obelisk project, so this commit just renames the thunk
someplace else.