Added a helper and error propagation to pipelines.
We'd been using withAsync to spawn worker threads for the logical
replication decoding pipeline, and only in some cases did we link the
thread to the main thread to detect errors; this silently dropped
asynchronous exceptions on the floor. withSingleWorker and
withSingleWorkerWatchdog are the helper functions.

withSingleWorkerWatchdog additionally provides an IO action that must be
run repeatedly, and a log message is being sent when it has been more
than 30s, to detect soft lockups. This mechanism can produce false
positives when there is little DB traffic.

We also had a soft lockup possible with the decoding step in the
pipeline requesting a new connection to read schema information every
time, and the pool running out of connections, which stopped progress
releasing connections.