Handle unexpected termination and fix watchdogs.
Switch to using the `race` combinator instead of `withAsync` to catch
unexpected but not exceptional termination of the worker thread.
withSingleWorker was using withAsync to spawn it's worker thread and
forwarding exceptions to the main thread; this led to a case where we
had a programming error missing a "forever" producing silent termination
of the worker thread.

Also correct the thread delay in withWatchdog to actually wait for
30 seconds and not 30 microseconds, and reinstate the watchdogs.